Rename heroeSlice to heroesSlice, fix typo Add more gap between the HeroCards
